refactor: Navbar 리팩토링
