+++ frontend/src/components/index.ts
@@ -1,2 +1,2 @@
-export { default as Navbar } from "./navbar/Navbar";
+export { default as NavbarSection } from "./navbar/NavbarSection";
 export { default as ProfileBox } from "./ProfileBox";
